fix: move BookingItem to BookingHelper
