+++ source/screens/BookingFormScreen.tsx
@@ -5,8 +5,7 @@ import styled from "styled-components/native";
 import AuthContext from "../store/AuthContext";
 import { createBooking, getTimeSlots } from "../services/BookingService";
 import { getReferenceCodeForUser } from "../helpers/ReferenceCode";
-import { formDataToQuestions } from "../helpers/BookingHelper";
-import { BookingItem } from "../helpers/MockBookingData";
+import { formDataToQuestions, BookingItem } from "../helpers/BookingHelper";
 import { getAdministratorsBySharedMailbox } from "../services/BookablesService";
 import BookingForm from "../containers/Form/BookingForm";
 import FormContext from "../store/FormContext";
